Feature/response helper

+++ src/main/java/com/github/glo2003/daos/ListingsDAO.java
@@ -1,29 +1,29 @@
 package com.github.glo2003.daos;
 
-import java.util.HashMap;
-import java.util.Map;
-
 import com.github.glo2003.helpers.MathHelper;
 import com.github.glo2003.models.Listing;
 
+import java.util.HashMap;
+import java.util.Map;
+
 public class ListingsDAO implements IDAO<Listing> {
 
-  private Map<Integer, Listing> listings = new HashMap<>();
+  private Map<Long, Listing> listings = new HashMap<>();
 
   @Override
-  public Listing get(int id) {
+  public Listing get(long id) {
     return listings.get(id);
   }
 
   @Override
-  public int save(Listing listing) {
+  public long save(Listing listing) {
     if (listings.containsValue(listing)) {
       // TODO return 0, the key or throw exception
     }
 
-    int id = MathHelper.randomInt();
+    long id = MathHelper.randomLong();
     while (listings.containsKey(id)) {
-      id = MathHelper.randomInt();
+      id = MathHelper.randomLong();
     }
     listings.put(id, listing);
     return id;
